requirements.txt
numpy==2.1.1
opencv-python==4.10.0.84
pillow==10.4.0
tkinterdnd2==0.4.2
